Fix bug in elnode--mapper-find that always found a match.
Refactored the find loop into it's own function for easier testing.
Removed the 'finally return' clause in the loop because it would
always return the last mapping.  It should fail instead so elnode can
return a 404.

Additionally, all regexps should start with ^.  Otherwise,
elnode--mapper-find might find a partial match and return it.  This is
most problematic with '$' which will match any string.  To match the
root you must use '^$'