do not start a server when required or loaded
This commit removes elnode-do-init and the related initialization
machinery.  The elnode-do-init customization variable can not be set
until after elnode is loaded.  By this time the server has already been
started so setting the value of elnode-do-init has no effect.  Thus this
customization variable is useless and should be removed.

There are many valid uses of elnode which do *not* require the default
server to be launched (e.g., running any non-default server, or
requiring elnode for development).  See issues #45 and #80.

This commit also updates the documentation of the elnode-init function.